fixed stop script

+++ scripts/stop.sh
@@ -1,6 +1,13 @@
 #/bin/sh
 
 forever_pid=$(ps aux | grep [f]orever | tr -s " " | cut -d " " -f 2)
-kill $forever_pid
+if [ ! -z "$forever_pid" ]
+then
+    kill $forever_pid
+fi
 node_pid=$(ps aux | grep [n]ode | tr -s " " | cut -d " " -f 2)
-kill $node_pid
+if [ ! -z "$node_pid" ]
+then
+    kill $node_pid
+fi
+
